Move file ID extraction into FileCheck
This logic is concerned with TDR file IDs, so it belongs with the rest of
the file check rather than the directory monitoring code.
