+++ backend-checks/src/main/scala/uk/gov/nationalarchives/tdr/localaws/backendchecks/FileWatcher.scala
@@ -3,23 +3,19 @@ package uk.gov.nationalarchives.tdr.localaws.backendchecks
 import java.nio.file.StandardWatchEventKinds.ENTRY_CREATE
 import java.nio.file._
 import java.nio.file.attribute.BasicFileAttributes
-import java.util.UUID
 
 import uk.gov.nationalarchives.tdr.localaws.backendchecks.checks.FileCheck
 
 import scala.collection.mutable
 import scala.concurrent.ExecutionContext
 import scala.jdk.CollectionConverters._
-import scala.util.{Failure, Success, Try}
 
 class FileWatcher(parentDirectory: Path, fileCheck: FileCheck)(implicit executionContext: ExecutionContext) {
 
-  private val UuidLength = 36
+  private val watcher: WatchService = FileSystems.getDefault.newWatchService
+  private val initialPaths: Map[WatchKey, Path] = registerAll(parentDirectory)
 
-  private val watcher = FileSystems.getDefault.newWatchService
-  private val initialPaths = registerAll(parentDirectory)
-
-  def watchDirectory: Unit = monitorChanges(initialPaths)
+  def startWatching: Unit = monitorChanges(initialPaths)
 
   private def registerAll(start: Path): Map[WatchKey, Path] = {
     val pathsByKey = mutable.Map.empty[WatchKey, Path]
@@ -55,7 +51,7 @@ class FileWatcher(parentDirectory: Path, fileCheck: FileCheck)(implicit executio
 
         Files.walkFileTree(fullPath, new SimpleFileVisitor[Path] {
           override def visitFile(file: Path, attrs: BasicFileAttributes): FileVisitResult = {
-            runFileChecks(file)
+            fileCheck.checkPath(file)
 
             FileVisitResult.CONTINUE
           }
@@ -63,7 +59,7 @@ class FileWatcher(parentDirectory: Path, fileCheck: FileCheck)(implicit executio
 
         currentPaths ++ newPaths
       } else {
-        runFileChecks(fullPath)
+        fileCheck.checkPath(fullPath)
 
         currentPaths
       }
@@ -72,32 +68,5 @@ class FileWatcher(parentDirectory: Path, fileCheck: FileCheck)(implicit executio
     watchKey.reset()
 
     monitorChanges(updatedPaths)
-  }
-
-  private def runFileChecks(path: Path): Unit = {
-    extractFileId(path) match {
-      case Success(fileId) => {
-        // Log any errors returned by the file check, then ignore them to allow the watcher to keep running
-        fileCheck.check(fileId).recover(error => {
-          println(s"Error saving antivirus result for path '$path'", error)
-        })
-      }
-      case Failure(e: IllegalArgumentException) =>
-        println(s"Filename at path '$path' does not end in a UUID, so skipping file checks")
-      case Failure(e) =>
-        println(s"Error extracting file ID from path '$path'", e)
-    }
-  }
-
-  private def extractFileId(path: Path): Try[UUID] = {
-    val fileName = path.getFileName.toString
-    if (fileName.length < UuidLength) {
-      return Failure(new IllegalArgumentException(s"Path '$path' is too short to contain a valid UUID"))
-    }
-
-    // S3Ninja does not store S3 folders as filesystem folders. Instead it puts underscores between the folder
-    // name and the filename, so we have to extract the last 36 digits of the filename to get the filename
-    // that would have been saved to S3.
-    Try(UUID.fromString(path.getFileName.toString.takeRight(UuidLength)))
   }
 }
